Added stupid enough implementation for default group Fixed dummy plugin
