+++ requirements.txt
@@ -1,3 +1,4 @@
+blinker==1.3
 bz2file==0.98
 click==3.3
 Flask-Assets==0.10
